feat: add `updatePlayerInTablist` functionality for improved tablist updates across commands and LuckPerms events
