+++ surf-tab-velocity/src/main/kotlin/dev/slne/surf/tab/velocity/command/SurfTabCommand.kt
@@ -4,6 +4,8 @@ import dev.jorel.commandapi.kotlindsl.anyExecutor
 import dev.jorel.commandapi.kotlindsl.commandTree
 import dev.jorel.commandapi.kotlindsl.literalArgument
 import dev.slne.surf.surfapi.core.api.messages.adventure.sendText
+import dev.slne.surf.tab.velocity.plugin
+import dev.slne.surf.tab.velocity.service.tablistService
 import dev.slne.surf.tab.velocity.tablistConfiguration
 
 fun surfTabCommand() = commandTree("surftab") {
@@ -12,6 +14,10 @@ fun surfTabCommand() = commandTree("surftab") {
     literalArgument("reload") {
         anyExecutor { executor, _ ->
             tablistConfiguration.reload()
+
+            plugin.proxy.allPlayers.forEach {
+                tablistService.updatePlayerInTablist(it)
+            }
 
             executor.sendText {
                 appendPrefix()
